+++ tests/ui/tests/support/k8sclient/commands.ts
@@ -1,7 +1,7 @@
 import {ApiRuleConfig} from "./apiRule";
 import Chainable = Cypress.Chainable;
 
-export interface K8sClient {
+export interface Commands {
     createApiRule(cfg: ApiRuleConfig): Chainable<JQuery>
     createService(name: string, namespace: string): Chainable<JQuery>
     createNamespace(name: string): Chainable<JQuery>
