+++ tests/ui/tests/support/k8sclient/index.ts
@@ -1,6 +1,6 @@
 import "./namespace"
 import "./service"
 import "./apiRule";
-export {K8sClient} from "./k8sClient";
+export {Commands} from "./commands";
 export {KubernetesConfig, getK8sCurrentContext} from "./kubeconfig";
